app-rag-conversation/pyproject.toml
[project]
name = "intro-to-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "pdemeulenaer",email = "pdemeulenaer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "streamlit (>=1.43.2,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "huggingface-hub (>=0.29.3,<0.30.0)",
    "gradio (>=5.23.2,<6.0.0)",
]

[tool.poetry]
packages = [{include = "intro_to_rag", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
